udpates to routing
