+++ packages/client/src/index.js
@@ -6,7 +6,6 @@ import * as serviceWorker from './serviceWorker'
 
 ReactDOM.render(
   <React.StrictMode>
-    {' '}
     <Routes />
   </React.StrictMode>,
   document.getElementById('root')
