+++ packages/client/src/pages/index.js
@@ -1,2 +1,2 @@
 export { default as Map } from './map'
-export { Planet, PlanetProvider } from './planet'
+export { default as Planet } from './planet'
