+++ packages/client/src/pages/planet/index.js
@@ -1,38 +1,19 @@
 import React from 'react'
-// import styles from './planet.module.css'
+import { Route, Switch } from 'react-router-dom'
+import Planet from './planet.container'
+import { PlanetProvider } from './planet.contex'
 
-import {
-  usePlanetState,
-  changePlanet,
-  usePlanetDispatch,
-  PlanetProvider
-} from './planet.contex'
-
-function Planet () {
-  // useLogic function to abstract messy code below return
-  const { planetState, goToPlanet } = usePlanetLogic()
-
+function PlanetPage () {
   return (
-    <div className='Planet'>
-      <p>You are now on planet {planetState.name}!</p>
-      <button onClick={() => goToPlanet('Jupiter')}>Jupiter</button>
-      <button onClick={() => goToPlanet('Earth')}>Earth</button>
-    </div>
+    <PlanetProvider>
+      <Switch>
+        <Route exact path='/'>
+          <Planet />
+        </Route>
+        <Route path='/leaderboard'>{/* <PlanetLeaderboard /> */}</Route>
+      </Switch>
+    </PlanetProvider>
   )
 }
 
-// contain all of the component logic here
-// so the render and view code is
-// easier to reason about. Export what you
-// need into the main function.
-
-function usePlanetLogic () {
-  const planetDispatch = usePlanetDispatch()
-  const planetState = usePlanetState()
-
-  const goToPlanet = name => changePlanet(planetDispatch, name)
-  return { planetState, goToPlanet }
-}
-
-// Wrap the Component in it's context provider
-export { Planet, PlanetProvider }
+export default PlanetPage
